feat(scraper): skip specials
this successfully skips the specials, but introduces a new bug.

I incorrectly assumed calculating the series letter as the first
alphabetic character in the title would work. but nay, some episodes
start with 'The'.

Also, the "Everything, Etc." markup is totally jacked up in wikipedia.
parsing that guy is gonna be a nightmare.